+++ src/ft_strlcat.c
@@ -13,9 +13,9 @@
 
 size_t	ft_strlcat(char *dst, const char *src, size_t size)
 {
-	int	i;
-	int	j;
-	int	count;
+	size_t	i;
+	size_t	j;
+	size_t	count;
 
 	if (!dst || !src)
 		return (0);
